Extract AbstractConnection class
